BackEnd/src/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
dj-rest-auth==6.0.0
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
